dumping BDD 1

+++ app/src/main/java/as/swarmapp/lighttracker/Imprevus.java
@@ -5,7 +5,7 @@ package as.swarmapp.lighttracker;
  */
 public class Imprevus {
 
-    private final static long UN = 0b1;
+    private final static int UN = 0b1;
     public final static long INCONNUE               = UN << 1;
     public final static long E_BAD_REQUEST          = UN << 2;
     public final static long E_SERVICE_UNAVAILABLE  = UN << 3;
